+++ packages/studio-shared/src/index.ts
@@ -64,14 +64,8 @@ export {
 	BrowserDownloadState,
 	BrowserProgressLog,
 	BundlingState,
-	ClientRenderJob,
-	ClientRenderJobProgress,
-	ClientRenderMetadata,
-	ClientStillRenderJob,
-	ClientVideoRenderJob,
 	CopyingState,
 	DownloadProgress,
-	GetBlobCallback,
 	JobProgressCallback,
 	RenderJob,
 	RenderJobWithCleanup,
